feat: add activeCycleId state
We need to memorize the active cycle in the application for use in the history page and also to identify what cycle is running.
So we create a state that when we start the counter, we set this state to the same id of the cycle. Then, we find who is the active by using a find to search who has the cycle.id equals to activeCycleId

This is a better approach because if  a boolean variable were created to store the active state, we would always have to iterate through the array and keep activating and deactivating cycles.